♻️ refactor(sync): migrate collection_books to descriptor catalog
